requirements.txt
# Core
python-dotenv
pyyaml
requests
agent_tooling
openai
setuptools
keyring
keyrings.alt

# MCP
mcp[cli]

# Gateway server
fastapi
uvicorn[standard]
aiosqlite
aiohttp

# External integrations
neo4j
py2neo
dateparser

# Discord bot
discord.py[voice]
PyNaCl

# Telegram bot
python-telegram-bot[job-queue]>=21.0

# Scheduler
apscheduler>=3.10

# Browser automation
playwright>=1.40.0

# Voice server (STT + TTS)
faster-whisper>=1.0.0
kokoro>=0.9.0
f5-tts>=1.0.0
torchcodec
torch
soundfile
numpy
